LayoutTests/imported/w3c/web-platform-tests/user-timing/test_user_timing_clear_measures-expected.txt

Description

This test validates that the performance.clearMeasures() method is working properly. This test creates the following measures to test this method:

"measure1"
"measure2"
After creating each measure, performance.clearMeasures() is called three times. First, it is provided with a name of "measureUndefined", a non-existent measure, which shouldn't change the state of the Performance Timeline. Next, it is provided with a name of "measure2", after which, this measure should no longer be present in the Performance Timeline. Finally, performance.clearMeasures() is called without any name provided. After this call, no measures should be present in the Performance Timeline. The state of the Performance Timeline is tested with the performance.getEntriesByType() and performance.getEntries() methods.

PASS window.performance is defined 
PASS Two measures have been created for this test. 
PASS After a call to window.performance.clearMeasures("measureUndefined"), where "measureUndefined" is a non-existent measure, window.performance.getEntriesByName("measure1") returns an object containing the "measure1" measure. 
PASS After a call to window.performance.clearMeasures("measureUndefined"), where "measureUndefined" is a non-existent measure, window.performance.getEntriesByName("measure2") returns an object containing the "measure2" measure. 
PASS After a call to window.performance.clearMeasures("measure1"), window.performance.getEntriesByName("measure1") returns an empty object. 
PASS After a call to window.performance.clearMeasures("measure1"), window.performance.getEntriesByName("measure2") returns an object containing the "measure2" measure. 
PASS After a call to window.performance.clearMeasures(), window.performance.getEntriesByType("measure") returns an empty object. 
